[fix] tests

+++ include/map/find.h
@@ -11,7 +11,7 @@ struct Unit;
 /*-- Map Usable -- */
 // move: find canEquip including on moveable tiles or not
 // Note: need entities/world, to find defendants
-void  Map_canEquip(struct Map *map, tnecs_entity ent, b32 move, int archetype);
+void  Map_canEquip(struct Map *map, tnecs_entity ent, b32 move, i64 archetype);
 
 /* -- Find at tile -- */
 tnecs_entity Map_Find_Door_Ent(     struct Map *map, i32 x, i32 y);
